[BUGFIX] Ensure the legacy script loads correctly
The AMD module shim was incorrect, causing the legacy IE script not to
load. This failed even in Chrome, but we didn't see the failure since
the script was never loaded and we haven't set up cross browser testing
yet. This PR fixes the issue and adds tests that ensure the legacy
polyfill loads correctly in Chrome at least.